feat: add sanitize module
