+++ index.html
@@ -19,6 +19,7 @@
         <div id="preview"></div>
       </div>
     </div>
+    <script src="sanitize.js"></script>
     <script src="markdown_editor.js"></script>
     <script src="codeBlockSyntax_java.js"></script>
     <script>
