+++ sanitize.js
@@ -1,0 +1,16 @@
+function sanitize(str) {
+  return str
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;')
+    .replace(/'/g, '&#39;');
+}
+
+if (typeof window !== 'undefined') {
+  window.sanitize = sanitize;
+}
+
+if (typeof module !== 'undefined') {
+  module.exports = { sanitize };
+}
